重构: 添加了 Github 仓库自动化使用的专用目录，并将部分内容放入此文件夹

+++ src/main/java/net/minecraft/world/level/levelgen/synth/SimplexNoise.java
@@ -61,7 +61,7 @@ public class SimplexNoise {
 
     protected static double dot(final int[] g, final double x, final double y, final double z) {
         if (isBedrockMode()) {
-            return g[0] * (float)x + g[1] * (float)y + g[2] * (float)z;
+            return (float)(g[0] * (float)x + g[1] * (float)y + g[2] * (float)z);
         }
         return g[0] * x + g[1] * y + g[2] * z;
     }
